Move Blazer generator to rolemodel:blazer
