+++ lib/generators/rolemodel/blazer/templates/app/assets/stylesheets/blazer.css
