+++ lib/generators/rolemodel/blazer/templates/app/assets/stylesheets/selectize.css
