+++ lib/generators/rolemodel/blazer/templates/app/controllers/reports/dashboards_controller.rb
