+++ lib/generators/rolemodel/blazer/templates/app/controllers/reports/queries_controller.rb
