+++ lib/generators/rolemodel/blazer/templates/app/views/reports/dashboards/index.html.slim
