+++ lib/generators/rolemodel/blazer/templates/app/views/reports/dashboards/show.html.slim
