+++ lib/generators/rolemodel/blazer/templates/config/initializers/blazer.rb
