+++ lib/generators/rolemodel/blazer/templates/lib/blazer_extensions/data_source.rb
