+++ lib/generators/rolemodel/blazer/templates/lib/tasks/reports.rake
