+++ lib/generators/rolemodel/blazer/templates/spec/factories/blazer_dashboards.rb
