+++ lib/generators/rolemodel/blazer/templates/spec/factories/blazer_queries.rb
