+++ lib/generators/rolemodel/blazer/templates/spec/system/reporting_spec.rb
